make it easier for extern parts to use the level.cc code
The external projects need to do some level handling. This can all be done with the
level.cc module but this module is a bit connected to other modules. So I introduced
the switch CREATOR that needs to be defined for these modules and that removes all the
functions that reference functions from other modules. 

You now define CREATOR and include the file level.cc. Thats nearly it. For details
look at the mission creation code in the levels project